Refactor the RV32 & RV64 assembly code for handling data allocation
This change refactors the way RV64 and RV32 assembly code handle data allocation in the emulator. The original data allocation procedure was overly complicated and hard to debug. The updated implementation simplifies data allocation by processing data emissions and initialization separately.

The updated procedure first checks the type of each line in the assembly file to figure out what allocations need to be done. If a line is identified as a data emission, it will create an object to store information about the emission. If it's a data initialization line, it will create a similar object but also include the initial value to be stored at that address.

This refactoring should make the data allocation process easier to understand and maintain. It follows best programming practices and improves the structure of code which should help for future feature additions and enhancements to the emulator.