Update SVG code for edit icon
The SVG code for the edit icon located at src/jsMain/resources/benicons/edit.svg was updated. The new SVG code appears to have an increased indentation for better readability and additional/modified elements, indicating changes to the icon itself. The changes were made for better visual appeal and to make the icon align better with the overall visual theme of the application.